Optimize bundling
