+++ src/util/fire-event.ts
@@ -18,18 +18,3 @@ export const fireEvent = <Event extends keyof HTMLElementEventMap>(
   });
   eventTarget.dispatchEvent(event);
 };
-
-export const hexFormatter = (bytes: number[] | Uint8Array) =>
-  "[" + bytes.map((value) => toHex(value) as any).join(", ") + "]";
-
-export const toHex = (value: number, size = 2) => {
-  let hex = value.toString(16).toUpperCase();
-  if (hex.startsWith("-")) {
-    return "-0x" + hex.substring(1).padStart(size, "0");
-  } else {
-    return "0x" + hex.padStart(size, "0");
-  }
-};
-
-export const sleep = (ms: number) =>
-  new Promise((resolve) => setTimeout(resolve, ms));
