+++ src/util/hex-formatter.ts
@@ -1,0 +1,4 @@
+import { toHex } from "./to-hex";
+
+export const hexFormatter = (bytes: number[] | Uint8Array) =>
+  "[" + bytes.map((value) => toHex(value) as any).join(", ") + "]";
